% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test}
\alias{test}
\title{Test Doc}
\usage{
test()
}
\value{
nothing
}
\description{
This is a test doc
\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}
\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\examples{
# no examples just fun
# just test
1+1
}
